requirements.txt
asgiref==3.5.2
black==22.3.0
certifi==2022.6.15
charset-normalizer==2.0.12
click==8.1.3
Django==4.0.5
djangorestframework==3.13.1
djongo==1.3.6
dnspython==2.2.1
idna==3.3
logdna==1.18.2
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.2
pycodestyle==2.8.0
pymongo==4.1.1
python-dotenv==0.20.0
pytz==2022.1
requests==2.28.0
sqlparse==0.2.4
tomli==2.0.1
typing_extensions==4.2.0
urllib3==1.26.9
